refactor(monitoring): extract state resolver
